Don't bundle string-split-html for client side
The package string-strip-html was being bundled for the client
side, even though it was only used on the server side (in a
loader function). Move the reference to it to a file with the
extension .server.ts to force it to be bundled only for the
server.
